Initial version
Add a Python script that can chop the original frames into 256 8x8 tiles.

This uses a bunch of heuristics to try to produce something that is visually
pleasing and not too far from the original while staying within the limits
of the GameBoy (assuming all tiles are loaded at all times; if we assume that
some tiles are only used in certain frames and that you can modify those
tiles from frame to frame, you can squeeze more out a higher number of
different tiles).